+++ R/methods_s3_plot.R
@@ -11,6 +11,12 @@
 #' @param make_gg logical. If \code{TRUE}, the plot is created using
 #' \code{ggplot2}. Otherwise, the plot is crated with base \code{R}.
 #' \code{TRUE} by default.
+#' @param plot_temp logical. Whether the temperature profile will
+#' be added to the plot. \code{FALSE} by default.
+#' @param label_y1 Label of the principal y-axis.
+#' @param label_y2 Label of the secondary y-axis.
+#' @param ylims Numeric vector of length 2 with the Limits of the
+#' y-axis. \code{NULL} by default (0, max_count).
 #'
 #' @return If \code{make_gg = FALSE}, the plot is created. Otherwise, an
 #'         an instance of \code{ggplot} is generated, printed and returned.
@@ -18,16 +24,65 @@
 #' @export
 #'
 #' @importFrom graphics plot
-#' @importFrom ggplot2 ggplot geom_line aes_string
+#' @importFrom ggplot2 ggplot geom_line aes_string sec_axis scale_y_continuous
+#' @importFrom ggplot2 ylab ylim
+#' @importFrom rlang .data
 #'
-plot.SimulInactivation <- function(x, y=NULL, ..., make_gg = TRUE) {
+plot.SimulInactivation <- function(x, y=NULL, ...,
+                                   make_gg = TRUE, plot_temp = FALSE,
+                                   label_y1 = "logN",
+                                   label_y2 = "Temperature",
+                                   ylims = NULL) {
+
+    # time <- NULL
+    # temperature <- NULL
 
     if (make_gg) {
 
-        p <- ggplot(x$simulation) +
-            geom_line(aes_string(x = "time", y = "logN"))
+        if (plot_temp) {
 
-        return(p)
+            min_time <- min(x$simulation$time)
+            max_time <- max(x$simulation$time)
+
+            max_count <- max(x$simulation$logN)
+
+            tt <- seq(min_time, max_time, length = 100)
+            min_temp <- min(x$temp_approximations$temp(tt))
+            max_temp <- max(x$temp_approximations$temp(tt))
+
+            slope <- (max_count)/(max_temp - min_temp)
+            intercept <- (min_temp * (max_count-0)/(max_temp - min_temp) - 0)
+
+            p <- x$simulation %>%
+                mutate(temperature = x$temp_approximations$temp(.data$time),
+                       fake_temp = .data$temperature * slope - intercept) %>%
+                ggplot() +
+                    geom_line(aes_string(x = "time", y = "logN"), linetype = 2) +
+                    geom_line(aes_string(x = "time", y = "fake_temp"))
+
+            if (is.null(ylims)) {
+                ylims <- c(0, max_count)
+            }
+
+            p + scale_y_continuous(limits = ylims,
+                                   name = label_y1,
+                                   sec.axis = sec_axis(~(. + intercept)/slope,
+                                                       name = label_y2))
+
+
+        } else {
+
+            p <- ggplot(x$simulation) +
+                geom_line(aes_string(x = "time", y = "logN")) +
+                ylab(label_y1)
+
+            if (!is.null(ylims)) {
+                p <- p + ylim(ylims)
+            }
+
+            p
+
+        }
 
     } else {
 
@@ -45,34 +100,52 @@ plot.SimulInactivation <- function(x, y=NULL, ..., make_gg = TRUE) {
 #' @param x the object of class \code{IsoFitInactivation} to plot.
 #' @param y ignored
 #' @param ... additional arguments passed to \code{plot}.
+#' @param make_gg logical. If \code{TRUE}, the plot is created using
+#' \code{ggplot2}. Otherwise, the plot is crated with base \code{R}.
+#' \code{FALSE} by default.
 #'
 #' @export
 #'
 #' @importFrom graphics plot lines title
+#' @importFrom dplyr mutate %>%
+#' @importFrom lazyeval interp
+#' @importFrom ggplot2 aes_string ggplot geom_point geom_line facet_wrap
+#' @importFrom stats predict
 #'
-plot.IsoFitInactivation <- function(x, y=NULL, ...) {
+plot.IsoFitInactivation <- function(x, y=NULL, ..., make_gg = FALSE) {
 
-    death_data <- x$data
-    model_data <- get_isothermal_model_data(x$model)
+    if (!make_gg) {
 
-    for (each_temp in unique(death_data$temp)) {
+        death_data <- x$data
+        model_data <- get_isothermal_model_data(x$model)
 
-        temp_indexes <- death_data$temp == each_temp
-        my_data <- death_data[temp_indexes, ]
+        for (each_temp in unique(death_data$temp)) {
 
-        # my_data <- subset(death_data, temp == each_temp)
+            temp_indexes <- death_data$temp == each_temp
+            my_data <- death_data[temp_indexes, ]
 
-        plot(log_diff ~ time, data = my_data, ...)
+            # my_data <- subset(death_data, temp == each_temp)
 
-        max_time <- max(my_data$time)
-        times <- seq(0, max_time, length= 100)
-        arguments_call <- c(list(time = times, temp = each_temp), x$parameters)
+            plot(log_diff ~ time, data = my_data, ...)
 
-        prediction <- do.call(model_data$prediction, arguments_call)
+            max_time <- max(my_data$time)
+            times <- seq(0, max_time, length= 100)
+            arguments_call <- c(list(time = times, temp = each_temp), x$parameters)
 
-        lines(times, prediction)
-        title(paste("Temperature:", each_temp))
+            prediction <- do.call(model_data$prediction, arguments_call)
 
+            lines(times, prediction)
+            title(paste("Temperature:", each_temp))
+
+        }
+
+    } else {
+        x$data %>%
+            mutate(prediction = predict(x$nls, newdata = x$data)) %>%
+            ggplot(aes_string(x = "time")) +
+                geom_point(aes_string(y = "log_diff")) +
+                geom_line(aes_string(y = "prediction"), linetype = 2) +
+                facet_wrap("temp", scales = "free")
     }
 }
 
@@ -89,6 +162,12 @@ plot.IsoFitInactivation <- function(x, y=NULL, ...) {
 #' @param make_gg logical. If \code{TRUE}, the plot is created using
 #' \code{ggplot2}. Otherwise, the plot is crated with base \code{R}.
 #' \code{TRUE} by default.
+#' @param plot_temp logical. Whether the temperature profile will
+#' be added to the plot. \code{FALSE} by default.
+#' @param label_y1 Label of the principal y-axis.
+#' @param label_y2 Label of the secondary y-axis.
+#' @param ylims Numeric vector of length 2 with the Limits of the
+#' y-axis. \code{NULL} by default (0, max_count).
 #'
 #' @return If \code{make_gg = FALSE}, the plot is created. Otherwise, an
 #'         an instance of \code{ggplot} is generated, printed and returned.
@@ -97,9 +176,13 @@ plot.IsoFitInactivation <- function(x, y=NULL, ...) {
 #'
 #' @importFrom graphics plot points
 #' @importFrom ggplot2 geom_point
-#' @importFrom ggplot2 aes_string
+#' @importFrom ggplot2 aes_string ylim
 #'
-plot.FitInactivation <- function(x, y=NULL, ..., make_gg = TRUE) {
+plot.FitInactivation <- function(x, y=NULL, ..., make_gg = TRUE,
+                                 plot_temp = FALSE,
+                                 label_y1 = "logN",
+                                 label_y2 = "Temperature",
+                                 ylims = NULL) {
 
     death_data <- x$data
 
@@ -110,9 +193,19 @@ plot.FitInactivation <- function(x, y=NULL, ..., make_gg = TRUE) {
 
     if (make_gg) {
 
-        pred_plot <- plot(x$best_prediction)
+        maxlogN0 <- max(c(x$data$logN,
+                          log10(x$best_prediction$model_parameters$N0)))
+
+        if (is.null(ylims)) {
+            ylims <- c(0, ceiling(maxlogN0))
+        }
+
+        pred_plot <- plot(x$best_prediction, plot_temp = plot_temp,
+                          label_y1 = label_y1, label_y2 = label_y2,
+                          ylims = ylims)
         p <- pred_plot +
-            geom_point(data = death_data, aes_string(x = "time", y = "logN"))
+            geom_point(data = death_data, aes_string(x = "time", y = "logN")) # +
+            # ylim(0, ceiling(maxlogN0))
         return(p)
 
     } else {
@@ -152,15 +245,29 @@ plot.FitInactivation <- function(x, y=NULL, ..., make_gg = TRUE) {
 #' @param make_gg logical. If \code{TRUE}, the plot is created using
 #' \code{ggplot2}. Otherwise, the plot is crated with base \code{R}.
 #' \code{TRUE} by default.
+#' @param plot_temp logical. Whether the temperature profile will
+#' be added to the plot. \code{FALSE} by default.
+#' @param label_y1 Label of the principal y-axis.
+#' @param label_y2 Label of the secondary y-axis.
+#' @param ylims Numeric vector of length 2 with the Limits of the
+#' y-axis. \code{NULL} by default (0, max_count).
 #'
 #' @return If \code{make_gg = FALSE}, the plot is created. Otherwise, an
 #'         an instance of \code{ggplot} is generated, printed and returned.
 #'
 #' @export
 #'
-plot.FitInactivationMCMC <- function(x, y=NULL, ..., make_gg = TRUE) {
+plot.FitInactivationMCMC <- function(x, y=NULL, ..., make_gg = TRUE,
+                                     plot_temp = FALSE,
+                                     label_y1 = "logN",
+                                     label_y2 = "Temperature",
+                                     ylims = NULL) {
 
-    plot.FitInactivation(x, make_gg = make_gg, ...)
+    plot.FitInactivation(x, make_gg = make_gg, plot_temp = plot_temp,
+                         label_y1 = label_y1,
+                         label_y2 = label_y2,
+                         ylims = ylims,
+                         ...)
 
 }
 
@@ -180,7 +287,7 @@ plot.FitInactivationMCMC <- function(x, y=NULL, ..., make_gg = TRUE) {
 #'
 #' @importFrom ggplot2 ggplot aes_string
 #' @importFrom ggplot2 geom_line
-#' @importFrom ggplot2 geom_ribbon
+#' @importFrom ggplot2 geom_ribbon ylab xlab
 #'
 #' @param x the object of class \code{PredInactivationMCMC} to plot.
 #' @param y ignored
@@ -213,7 +320,8 @@ plot.PredInactivationMCMC <- function(x, y=NULL, ..., make_gg = TRUE) {
         p <- ggplot(x, aes_string(x = "times")) +
             geom_ribbon(aes_string(ymax = "upper", ymin = "lower"), alpha = 0.6, fill = "#56B4E9") +
             geom_line(aes_string(y = "mean"), linetype = 2, colour = "darkblue") +
-            geom_line(aes_string(y = "median"), linetype = 3, colour = "darkblue")
+            geom_line(aes_string(y = "median"), linetype = 3, colour = "darkblue") +
+            ylab("logN") + xlab("time")
         return(p)
 
     } else {
@@ -242,3 +350,7 @@ plot.PredInactivationMCMC <- function(x, y=NULL, ..., make_gg = TRUE) {
         }
     }
 }
+
+
+
+